requirements.txt
altair==5.1.0
ansi2html==1.8.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
category-encoders==2.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deprecation==2.1.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.2.5
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
ipython==8.14.0
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kaleido==0.2.1
kiwisolver==1.4.5
lightgbm==4.0.0
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
nbformat==5.9.2
nest-asyncio==1.5.7
numba==0.57.1
numpy==1.23.5
orjson==3.9.5
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.16.1
plotly-resampler==0.9.1
pmdarima==2.0.3
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycaret==3.0.4
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyod==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rich==13.5.2
rpds-py==0.10.0
schemdraw==0.15
scikit-base==0.5.1
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.11.2
six==1.16.0
sktime==0.22.0
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.26.0
tbats==1.1.3
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trace-updater==0.0.9.1
traitlets==5.9.0
tsdownsample==0.1.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wurlitzer==3.0.3
xxhash==3.3.0
yellowbrick==1.5
zipp==3.16.2
